requirements.txt
fastapi
uvicorn
minio
pika
boto3
numpy
pdfplumber
#textract
pdf2image
#easyocr
pdfminer.six[image]
docx2python
pypandoc
python-docx
PyMuPDF
redis
chonkie[st]
qdrant_client